refactor: move component into webchat chat area
