+++ packages/botonic-react/src/webchat/webchat-chat-area/styles.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const StyledWebchatChatChatArea = styled.div<{ height: number }>`
+  display: inherit;
+  flex-direction: inherit;
+  height: ${props => props.height}px;
+  width: inherit;
+  overflow: inherit;
+`
